Added in offset collect code to main method
